Vassal and confirm menu + new landmarks roles

+++ src/main/java/org/leralix/tan/gui/user/property/PropertyMenus.java
@@ -6,8 +6,8 @@ import org.bukkit.event.inventory.ClickType;
 import org.bukkit.inventory.ItemStack;
 import org.leralix.tan.dataclass.PropertyData;
 import org.leralix.tan.gui.BasicGui;
+import org.leralix.tan.gui.common.ConfirmMenu;
 import org.leralix.tan.gui.cosmetic.IconKey;
-import org.leralix.tan.gui.legacy.PlayerGUI;
 import org.leralix.tan.lang.FilledLang;
 import org.leralix.tan.lang.Lang;
 import org.leralix.tan.listeners.chat.PlayerChatListenerStorage;
@@ -148,13 +148,16 @@ public abstract class PropertyMenus extends BasicGui {
                 .setName(Lang.GUI_PROPERTY_DELETE_PROPERTY.get(langType))
                 .setClickToAcceptMessage(Lang.GUI_GENERIC_CLICK_TO_PROCEED)
                 .setAction(event ->
-                        PlayerGUI.openConfirmMenu(player, Lang.GUI_PROPERTY_DELETE_PROPERTY_CONFIRM.get(langType, propertyData.getName()),
-                                p -> {
+                        new ConfirmMenu(
+                                player,
+                                Lang.GUI_PROPERTY_DELETE_PROPERTY_CONFIRM.get(propertyData.getName()),
+                                () -> {
                                     propertyData.delete();
                                     player.closeInventory();
                                 },
-                                p -> open()
-                        ))
+                                this::open
+                        )
+                )
                 .asGuiItem(player, langType);
     }
